refactor: rename from List to AnimeList
