+++ android/features/animeList/src/main/kotlin/jp/co/yumemi/features/animeList/AnimeListItem.kt
@@ -25,7 +25,7 @@ import jp.co.yumemi.core.primitives.Spacing
 import jp.co.yumemi.core.R
 
 @Composable
-internal fun AnimeListCard(
+internal fun AnimeListItem(
     title: String,
     seasonName: String,
     onClick: () -> Unit,
@@ -72,9 +72,9 @@ internal fun AnimeListCard(
 
 @Preview
 @Composable
-private fun AnimeListCardPreview() {
+private fun AnimeListItemPreview() {
     Surface {
-        AnimeListCard(
+        AnimeListItem(
             image = R.drawable.img_no_image,
             title = "Title Japanese",
             seasonName = "Season Name",
