+++ kmp/presentation/sample-shared/src/commonMain/kotlin/animeList/AnimeListState.kt
@@ -1,28 +1,28 @@
-package jp.co.yumemi.list
+package jp.co.yumemi.animeList
 
 import jp.co.yumemi.domain.core.DomainError
-import jp.co.yumemi.domain.entities.ListCardEntity
+import jp.co.yumemi.domain.entities.AnimeListItemEntity
 import jp.co.yumemi.utils.Parcelable
 import jp.co.yumemi.utils.Parcelize
 import tech.fika.macaron.core.contract.State
 
-sealed class ListState : State, Parcelable {
+sealed class AnimeListState : State, Parcelable {
     /** 初期状態 */
     @Parcelize
-    data object Initial : ListState()
+    data object Initial : AnimeListState()
 
     /** 初回ローディング */
     @Parcelize
-    data object Loading : ListState()
+    data object Loading : AnimeListState()
 
     /** 通常 */
-    sealed class Stable : ListState() {
-        abstract val list: List<ListCardEntity>
+    sealed class Stable : AnimeListState() {
+        abstract val animeList: List<AnimeListItemEntity>
 
         /** 画面 */
         @Parcelize
         data class Initial(
-            override val list: List<ListCardEntity>,
+            override val animeList: List<AnimeListItemEntity>,
         ) : Stable()
     }
 
@@ -30,5 +30,5 @@ sealed class ListState : State, Parcelable {
     @Parcelize
     data class Error(
         val error: DomainError
-    ) : ListState()
+    ) : AnimeListState()
 }
